Added the STARTTLS connection mode
Adds the STARTTLS connection mode in function _server_authenticate. It's
use by Microsoft Hotmail and Exchange servers. I use the
MAIN_MAIL_EMAIL_STARTTLS const introduced in admins/mails.php retrieved
from the $conf object which is not used anywhere else in the file so it
may not be the right way to do it. Maybe this value should be passed in
constructor but it would then imply more changes around multiple files
calling smtps class.