Adds a STARTTLS option
Adds a STARTTLS authentification option. This is used by Microsoft
Hotmail and Exchange. This commit includes a new constant named
MAIN_MAIL_EMAIL_STARTTLS, a view and edit view for it and the edit
action needed to modify value in DB. It also updates the js function
initFields to make it behave like the standard TLS option and adds two
jQuery function respectively to set TSL off when STARTTLS is selected
and to set STARTTLS off when TLS is selected. Unselecting one of those
won"t select the other.