Fix display issue on step 4 of import in import.php. The issue was only for custom module import as the entity was initialised with the icone name i.e. filename@modulename.
Then the arrays entitytoicon and entitytolang where used. But by
definition custom modules do not have entries in these arrays.

So changed the assignement to proper object property when $entity is not
in these 2 tables. This is not changing behaviour of default object.

I have tested both with my custom objet and product object. So it should
be fine now.