[Qual] Optimize language file loading
Using fscanf() and the proper format is about twice as fast as
using fgets() followed by explode() and trim().

This is a small optimization per se but since this code is
called often it yields about 33% rendering time improvement on
the homepage where about 50% of the time is spent on loading the
translation files.