Add: Overload styles files from a current folder of module & theme
My goal is to overload css files only for the current module and without changing the source code of the module but just by adding a folder style in the module folder. Then just place the css files to overload in this folder they will be automatically added in the head.

The same method is applied to the current theme.

To test, for example, place a custom.css file in the societe/style/custom.css folder and check the head of a company module page.