Update admin.lang
Note that 80% of current Dolibarr users use the term "Accounting" to say "invoicing" (Definition of work accounting is not same for end users and for book keeping people). So to avoid confusion for end users and not having letting think they need this module because they ant to make accounting (with their definition), we must to be clear in name (description is not enough). So i rename "Accounting" into "Accounting (double entry)". So this is clear it is accounting with the definition of a book keeper.