allow to remove one custom referent in the project.
Allow to remove one custom referent from the project.

When we develop one module that supports the custom referent in a
project and if you don't follow the same naming convention as in the
other table (in french). You cannot remove the element.

This fix support now the property 'project_field'. And thus set this
column to null.

[see: #11881]