Add finalize feature to share contacts between thirdparties
This completes the feature activated by
$conf->global->MAIN_SUPPORT_SHARED_CONTACT_BETWEEN_THIRDPARTIES.
Modified CommonObject::list_contact() (added param 'self' to $source
param) to also get contats owned by a thirdparty itself (this permit to
get in the same list the extrnal contacts linked to a thirdparty
(llx_element_contact) and the contact owned by a thirdparty
(llx_societe_contacts)).  Also modified in Societe class, the functions
to get the contacts to be able to return the external contacts
(llx_element_contact) in addition of the thirdparty contacts
(llx_societe_contacts) (added '$getexternal' param for these functions).