Fix: Bug when delete payment when already deleted in account
Related to : 
https://github.com/Dolibarr/dolibarr/blame/11.0/htdocs/compta/bank/class/account.class.php#L1976

Mistake between id / rowid. I prefere add the fix here (but a fix in account class can be also a good fix). 
And best fix may use setter to do it (but to complicated : maybe latter ? i may check it in september if you want in all CommonObject properties)