FIX BUG - Subscrition type may not be confirmed
Currently, the subscription view displays by default the adherent's current subscription type if no type has been registered in the past.

If an update was made from an earlier version of Dolibarr (up to v10), the subscription type was not recorded in the tablellx_subscritipion. There is therefore nothing to display.

This procedure is risky because it misleads the user into believing that the information is confirmed. In my opinion, it is better not to display anything as a type of subscription when the information does not exist.

In certain circumstances, for legal and insurance matters, the history is particularly important. You should know if the adherents were covered several years back.

If the information is missing on the platform, the user will know that he will have to retrieve this information otherwise, but if he sees information, he risks saying that it is true.

It's far too risky.