FIX #14834 SQL query use INNER JOIN not comma
Not sure if both problem have the same origin, but the SELECT SQL is wrong cause "llx_facture_fourn" and "llx_societe_commerciaux" should be JOIN formally (not with a comma) in a multiples left joint SQL query.

See : https://dev.mysql.com/doc/refman/8.0/en/join.html

* INNER JOIN and , (comma) are semantically equivalent in the absence of a join condition: both produce a Cartesian product between the specified tables (that is, each and every row in the first table is joined to each and every row in the second table).

However, the precedence of the comma operator is less than that of INNER JOIN, CROSS JOIN, LEFT JOIN, and so on. If you mix comma joins with the other join types when there is a join condition, an error of the form Unknown column 'col_name' in 'on clause' may occur. Information about dealing with this problem is given later in this section. *