CLOSE #15065 Put the product label in bold in the PDF templates if configured
The global configuration variable PDF_BOLD_PRODUCT_LABEL is checked to put the product label in bold in the PDF
templates.
This modification is based on the same behaviour for the product reference, found at line 1373 of the same file.

The products table of the invoices / commercial propositions is produced in the PDF library, hence we have to modify
this library to style the invoices.
Producing a new template would mean duplicating code from the library with the maintenace implications.