[FIX] sort order in birthday box
The current sort order is by date and not by day of month. This yields
the following unexpected results:

Given

* Person A, born on 1980-01-10
* Person B, born on 1970-01-30

I expect Person A to be listed first because their birthday is before
that of Person B _in January_. But without this fix, Person B would show
up first because 1970 is before 1980.