Fix use login for User homedir in LDAP
posixAccount objectclass (intended to be used with the uid/gid/homedir fields) requires a homedir.
Always set a homedir in LDAP for the User class even if user does not have a firstname by using the login instead.
Additionally on Linux the login is typically used for the home directory rather than the first name.
It also avoids having accentuated or other special characters (commonly found in names) in the home directory path.