Fix #18895 Fix negative stock correction
If people use a negative number as input, I think we can assume they want to delete units. It will now set the movement to 1 if the number is negative and 'add' was selected.

For each negative number we now multiply this number by -1, to make it a positive number again.