Fix PR #18845 and #18843 (error on src_objet_type from emc files)
The PR #18845 and #18843 modify `CommonObject::commonGenerateDocument()`
and emc auto index page to use a `'@'.$object->module` suffix appened to `
`@object->table_element` when object has `$object->module` property. It's a
mistake (when originating object is referenced in stock movement table,
the module part is indicated after a `@` suffix and i thought it was a convention for
specify module part but in here `_` must be used).

Before the PR #18445 and #18443, `$object->table_element` was used to
set `src_object_type` in `llx_ecm_file`, normally the tables are named
like `$object->module.'_'.$object->element` but for random reason it
possible to name them otherwise (objects has specific `table_element`
property for that). So in CommonObject::commonGenerateDocument(), if the
object has a property `$this->module` use
`$this->module.'_'.$this->element` seem more reliable than use
`$this->table_element`.