FIX : Invoices Statistics corresponding to the fiscal year dates
FIX #19646 Invoices Statistics corresponding to fiscal year

The update was conducted on the file htdocs\compta\facture\class\facturestats.class.php.
The target is to display statistics and charts of invoices based on fiscal year and not calendar year.
For that purpose we used the configuration value of SOCIETE_FISCAL_MONTH_START to calculate date interval values.