FIX Solving non-blocking bug in user login
When using the function check_user_password_dolibarr to authenticate a user, a call is made to the function dol_verifyHash which requires a pre-calculated hash. This hash is not available with programmatically-created never-used accounts. Hence, in suce cases we have the following warning:
Notice: Trying to access array offset on value of type null in .../htdocs/core/lib/security.lib.php on line 156

This fix solves that bug by avoiding that unnecessary call