Update Segment.php
Hi Eldy.

You are right, {object_date} has the inmutable invoice date but the fact is that it is not available in the 'segment' file (or I was not able to find it) so I created the new function in the 'odf' file to retrieve it. It is the only modification to that file.

In the meantime you also are right that __CURRENTDAY__ should give the value relative to the current date to be consistent, not the one of the invoice. As I want, and need, the one on the invoice I have created a new block where 'CURRENT' has been changed by 'INVOICE' to build equivalents to the others (now relative to the current date) but relative to the invoice date.

Hope this clarifies the used tags and validate the changes.