Stock at date estimated value calculation
The calculation was done in the sql request (pmp * reel or pmp * stock depending if a warehouse has been selectionned).
But the real stock at date is calculated below this request (from l. 509 to l. 534). 
The estimated value displayed  (l. 567) is not correct anymore and should use the calculated stock instead.