NEW inc.php: handle parameters from argv
This commit adds support for --help and --config to provide a different
config file when running upgrade.php from commandline. The argv array
is patched afterwards as if the options were never given so that it
stays transparent for the code path that are not aware of those
arguments.

The intention behind is incrementally move the usage of argc/argv to
this location and force help/usage from there, in particular for
upgrading.

The rationale for the --config addition is to be able to provide a
different path for conf.php when multiple dolibarr instance are using
the same readonly htdocs folder, which is already possible by modifying
the `include_path` from PHP for the htdocs/ directory but not for the
htdocs/install directory since relative paths are used to fetch the
config file. Since the use-case is to upgrade/migrate a Dolibarr
instance from CLI, it makes sense to be able to select for which
instance (and database parameters) the upgrade should take place.