NEW test/phpunit: skip tests when requirement fails
Skip tests instead of die() when the requirement for the test to run
(like for instance module requirement) is not matching what the setup
has.

On a normal phpunit run, this leads the tests to be reported as
"skipped" and would still exit(0) after the testsuite has run, but
they can stil be made to fail the testsuite through --fail-on-skipped
phpunit option.

This allows running the testsuite with different configuration, while
knowing why the test was skipped for every test at once if the proper
configuration is wanted in the end.