FIX: Update mouvementstock.class.php (#25396)
* FIX: Update mouvementstock.class.php

If first character change to upper case, some external modules don't show its origin in mouvements.

* Update mouvementstock.class.php

More elegant.
We don't need to change to upper case (line 978) and then to lower case (line 980). It is only necessary when creating a new class (line 983)
The module name and url are in lower case.