Fix #26005 - link for EU VAT ID validation changed
Note: The validation link is defined as a Transifex translation key, but same value/URL has now to be incorporated in all languages. Therefore a change just in Transifex and just for the German version would not make sense, same value needs to be incorporated in all languages, therefore changed via pull request.

In theory, if EU offered different pages for different member countries, it could be adopted. Currently it's the same link for everyone.