NEW Generic doc template for donations
Can be used to create written acknowledgments to substantiate a charitable
contribution in the United States, but it is not specific to the United
States.

Multi-lingual.

If the global setting DONATION_NOTE_PUBLIC is greater than 0, the public
note for the donation is also included. This can be used in the United
States to notate if a donation is not entirely deductible.