Update stockatdate.php (#26481)
As commented in #23782 this update is to clean and clarify the code for future evolution of stocktodate.
In fact the calculated value in the SQL request is the current stock value and not the estimated stock value at date value.
Therefore I changed in SQL request the estimatedvalue to currentvalue and reused as a variable for a better understanding and for future use (other PR) as it will be interesting to get a comparison between stockatdate value to currentstock value. Also some others changes are required to improve the page speed.