Add SumUp connector compatibility with SumUp app on iOS
https://github.com/sumup/sumup-android-api
VS https://github.com/sumup/sumup-ios-url-scheme

The amount field is deprecated on the Android API but not on the iOS API. The iOS API does not support the total field, only the Android API does support the total field. Hopefully, the Android API only supports the amount field that is only partially deprecated on the Android API. For maximum compatibility with both Android/iOS SumUp app, best to keep this field as "amount" rather than "total".