Fix: pre-commit & codespell configuration.
# Fix: pre-commit & codespell configuration.

When codespell is run from the CLI it gets its configuration from
pyproject.toml or .codespellrc .
`pyproject.toml` was used because other tools also use that while
.codespellrc is specific.

A typo was fixed in the pre-commit configuration.

The end of file fixer was limited to yaml files, but that was mainly
to fix the yamllint errors that were reported in ci.
This hook can be applied on the files this normally applies to
without requiring it (not in ci, only local).