FIX Https detection
theme/eldy/manifest.json.php was returning http:// links instead of the
https link defined in the configuration.
I found that this is because I am behide a proxy and that isHTTPS does
the detection as expected.

Therefore that function is used to detect https which results in the correct
urls.

Detected on my 18.0.2 installation