Qual: Fix apstats & add phan (#28127)
# Qual: Fix apstats & add phan

I fixed several issues that I ran into when running apstats.php
locally to test the phan integration.

So this commit only proposes the updates to apstats.php.

I added a test that disables phan when its configurations is not
found so it should be fine to integration in develop.