Qual: Apply php-cs-fixer to make only manual changes stand-out (#28300)
* Qual: Apply php-cs-fixer to make only manual changes stand-out

# Qual: Apply php-cs-fixer to make only manual changes stand-out

A few changes will be proposed in the future to these files.  This already commits
the esthetic changes using php-cs-fixes.

* Qual: Apply php-cs-fixer to make only manual changes stand-out

# Qual: Apply php-cs-fixer to make only manual changes stand-out

A few changes will be proposed in the future to these files.  This already commits
the esthetic changes using php-cs-fixes.

* Qual: Apply php-cs-fixer to make only manual changes stand-out

# Qual: Apply php-cs-fixer to make only manual changes stand-out

A few changes will be proposed in the future to these files.  This already commits
the esthetic changes using php-cs-fixes.