Fix typing for showOutputFieldForObject
# Fix typing for showOutputFieldForObject

Type hint that $value can also be a list of strings.
Also updated the "else-if" sequence so that a static tool checker can
still appropriately detect when $value is a string and when it is
a list of strings so that it provides correct feedback.
Added an error message to the syslog in case we got an array but not
"array" as the value for "$type"