FIX: online signature partly fails if commercial proposal numbering contains a mask that’s includes a space (#28608)  (#28742)
* FIX: online signature (#28608) Update signature.lib.php

If Commercial proposal numbering contains a mask that’s includes a space, that will affect the result of the online signature form. When generating the online signature form, the code in signature.lib.php line 77, will remove any existing space in the Commercial proposal numbering. The result is that the online signature form will miss data in field “Third-party” and “Amount”.

* FIX: online signature (#28608) Update payments.lib.php

If Commercial proposal numbering contains a mask that’s includes a space, that will affect the result of the online signature form. When generating the online signature form, the code in core/lib/payments.lib.ph, will remove any existing space in the Commercial proposal numbering.