Update bonprelevement.class.php
Problem


When creating the file sepa ( Direct Debit File) if the amount ends in 0, such as 50.60, the program generates the file with rounding 50.6.


Some banks, for example, La Caixa (CaixaBank) do not accept the direct debit file if the decimal number does not have two digits.


Solution:


What I propose is to create a configuration variable called SEPA_ROUND_TWO_ZERO that allows for a two-digit rounding even if it ends in 0. That is, if the amount is 50.6, the file will be generated with the amount 50.60.