Fix: content-type haeder missing for webhook (#29588)
Data sent by webhooks is encoded in JSON and sent using the CURLOPT_POSTFIELDS function in cURL. However, the Content-Type header is not explicitly specified. By default, cURL automatically sets the Content-Type header to application/x-www-form-urlencoded, which can cause issues with certain applications.
To avoid this, I explicitly set the Content-Type header to application/json. This ensures that the data is sent in the correct format and that applications receiving the data can process it correctly.