FIX update date_echeance of supplier invoices when we update invoice date in the past
When we update the date of supplier invoices in the past, the date_echeance was not updated if the new calculated date_echeance is before the current date_echeance.
The same goes if we update the payment term of supplier invoice, the date echeance is not updated if the new calculted date_echeance is before the current date_echeance.
With this PR, when we update the supplier invoice date or payment term, the date_echeance is updated even if the new calculated date_echeance
 is before the current date_echeance.