Allow tweaking of the index that MySQL use when searching products (#29934)
* Allow tweaking of the index that MySQL use when searching products

Sometimes when working with large dataset with many entities the MySQL engine does not pick-up the right index.
This is especially true when doing a search over the reference because the LIKE '%%' prevent the uk_product_label from being chosen.

This PR aims to improve this situation by doing the following :

- Adding a simple index on entity column.
- Adding an hidden constant to force the index being used when searching products.

Performances over a 8M products dataset shows a reduction of 2/3 execution time.

* Add missing global $conf

* Fix wrong usage of getDolGlobalString()

* Finalise fix

* Add more indexes

* Rework indexes