CLOSE #26384 Implement STATUS_OPENEXT_ALL for use of non-counting warehouses e.g. quarantine
This feature complements the need explained in #26384 and the implementation of the STATUS_OPENEXT_ALL warehouse type. With the changes a warehouse can be used e.g. as a quarantine warehouse or a temporary pre-QC warehouse. The status "Open (omitted from stock calculation)" keeps the definition vague enough as not to constrain the use of this status too much in the sense of "either quarantine or pre-QC".
For code-readability and convenience the change is using the load_stock and loadWarehouses implementation. As the option of warehouseclosed, warehouseopen and warehouseinternal is already introduced in the code we can piggyback ride with the introduction of the STATUS_OPENEXT_ALL onto the product list view, stock transfer, delivery selection etc.